trashpi/requirements.txt
# TrashPi - Standalone ruslaflokkari
# Raspberry Pi + Camera + LEDs + Speaker

tflite-runtime>=2.14.0
picamera2>=0.3.12
gpiozero>=2.0
RPi.GPIO>=0.7.1
pyttsx3>=2.90
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
